Added a label field to Address, removed it from Storage.
An Address can store a Label with an offset, but no base or index,
and it will be ancoded as a RIP-relative address.

Renamed float_sign_bit_label to float_minus_zero_label, and filled
with double values.

Removed some X64 methods that dealt with a separate Label and offset.

Float literals now return MEMORY storage with the same label.