Allow member access of rvalue records.
When looking up a member of a partial variable, the pivot value not only
cast off the partial type, but becomes an lvalue, so in the case of record
initializer functions the member also stays lvalue, so it can be made
uninitialized.

$ in record methods is an lvalue, so members also stay lvalues.

A member variable of a rvalue record is now not an lvalue, and will be
extracted from the STACK storage record.