Yield-s can return simple values in registers.
EvalScope no longer stores details, only a pointer to the YieldableValue,
which handles everything.

Yield and YieldValue directly use the YieldableValue, not the EvalScope.

YieldableValue only creates an actual variable for non-simple results,
otherwise the RAX/XMM0 registers will be used for the yielded value, which are
preserved by the unwinding process.