Exceptions are stored in RDX during unwinding.
Removed EXCEPTION_ADDRESS and RESULT_ALIAS_ADDRESS macros,
and the related hidden local variables from the function bodies.

Renamed PUSHF/POPF to PUSHFQ/POPFQ, since they operate on 64-bit values.

Added rflags printing to the dump runtime function. The register values
are passed by a pointer, because pretending that it's passed on the stack by
value was only a guess, and failed after adding one more field to it,
probably due to some alignment issue.

Prohibit using the fixed NULL Address in effective address calculations.

FunctionScope now dynamically allocates hidden variables for the result
alias (for returning complex values), and forwarded exceptions (for
forwarding an exception from a Code back to the caller).

CodeScopeValue uses RAX/XMM0 for saving return values instead of picking a
register, so it won't interfere with and exception in RDX. Also, it
clears RDX before finalizing.

Replaced EXCEPTION_ADDRESS usage with RDX everywhere. The value in RDX
can be safely dropped when an exception is caught, and regular execution
resumes. RDX is used as a qword, which, when positive, stores an
unsigned byte treenum, but when negative, a full qword yield index.

IsValue no longer needs a matched variable to keep track if an exception
came from the matching expression, or the then branch, instead it just maps
matching UNMATCHED exceptions to a special value.

Fixed TryValue not using a Code for its block.

FunctionDefinitionValue allocates a storage for forwarded exceptions
if the function has Code arguments.

Exception forwarding now happens inside such functions, the EvaluableValue
raises the user exception, the function stores it for forwarding and translates
it into CODE_BREAK, and when returning, raises the user exception again.