Replaced Args and Kwargs with plain containers, as nothing else is needed, since even keyword arguments are unordered.
Modified check_arguments because of this, now we check keyword arguments
in the arg_infos order, not the specified order. This is necessary for
controls where a variable is declared in one argument and used in another.

Collected struct-s in globals.h on the top.

PartialInfo is now a struct.

Removed the kludgy special handling in check_arguments, where a NO_TS
argument type could be used to make an argument optional. This was
wrong anyway. Now the rbtree class handles this case properly.