Added Scope leaving checks.
Added Scope::leave method, to be called after all typizations are complete
within the scope. Will be used for scoped imports/exports later.

Added Declaration::outer_scope_left method, to be called when the
outer scope's leave method is called. Currently only used in Scope
to make sure inner scopes are left before their outer scope is.

FunctionScope allows the body scope to be added after itself is left.
This is necessary, because explicitly scoped functions have their
definitions completed after the scope is left, so FunctionScope
must be left before the body scope is created. But this is fine.

Readded Role::complete_role, so it can leave its inner scope, which is
necessary to allow subsequent declarations with explicit roles put
declarations into it.

The check that Interface can only contain functions and Implementations
is moved from InterfaceDefinitionValue to InterfaceType.

The check that Implementation functions correctly implement the interface
functions is moved from ImplementationDefinitionValue to ImplementationType.

The addition of the record compare operation is moved from RecordDefinitionValue
to RecordType.

Type::set_name also sets the inner scope's name, so no need to do this
in the type definition values.

Added a bunch of scope leave method calls.

TypeDefinitionValue is now a tiny base class, so non-scoped definitions can
also subclass it.

Added ScopedTypeDefinitionValue class for the others.