Added import of specific identifiers.
Added NamedScope common base class of DataScope, ModuleScope, and
RootScope, implementing the naming, meta and extension scope usage, and
upward growing reservation.

ModuleScope now uses the root scope as its meta scope, so identifiers
will be automatically be looked up there if the local lookup fails.
This makes the root level outer scope hacking unnecessary, as adding
Modules to Root can be delayed until the module ordering.

Renamed ModuleImportScope to ImportScope, and it exports itself into
the current module's scope instead of the root scope, so imports can
shadow module level identifiers, as to be expected.

ImportScope does a lookup of a set of explicit identifiers, in addition
to the prefix based matching.

Global type names and module qualified names are only looked up in the
current module scope.