Reference to pointer conversion correctly arranges an unborrow.
Added a Scope * argument to a bunch of methods, such as Scope::lookup,
Declaration::match, Identifier::matcher, Type::lookup_inner,
Type::lookup_initializer, Type::lookup_matcher.

Reordered arguments of typematch, now the Scope * argument is mandatory,
and is needed for the automatic Ref to Ptr conversion.

Added Unborrow class that executes a decref in its finalizer.

Fixed ReferenceBorrowValue class to properly convert a Reference to a
Pointer, arranging an Unborrow in its scope.

Added a six-argument version of the make template function.

Renamed :weak to :borrow.