Added proper Lvalue container borrows.
VariableValue classes get a Scope constructor argument.

Added Type::borrow method, which is only implemented in PointerType,
and is used to find the heap object's reference counter, and pass it to
an Unborrow instance.

Removed Runtime::borrow_dummy_label, because it was just accidentally merged
into this branch.

Added borrowing to ContainerIndexValue, MapIndexValue, SliceIndexValue,
VariableValue classes.