Implemented rbtree cloning.
Added workaround for NosyValue cloning. NosyValue needs to allocate a new
FCB for the clone, and that needs the address of its NosyContainer. The
container_cow function makes sure it's loaded into RDX for that purpose.
But currently Weak* types don't share their Rbtree, so this can't be
triggered yet.

Set and Map {add,remove} operations now need an Lvalue pivot.

Fixed Queue cloning not copying the FRONT field.

Fixed rbtree_allocate not updating the LENGTH field if a vacancy was filled.