Added explicit scope enter method.
This is a preparation step for the actual multipass compilation.

Some basic types no longer pretend that their inner scope is their
metascope, and must be made an inner scope explicitly.

Array, Queue, and Float no longer create their own inner scope either.

Adding and removing declarations are only allowed in entered scopes.

Adding a declaration automatically treated as their scope was entered,
and removing them as the scope was left.

Scopes are created in left state, and must be entered explicitly.

ExportScope only pushes itself to the target scope when its outer scope
is entered.

Added lots of Scope::enter calls.

ScopedTypeDefinition only creates the type inner scope in complete_definition,
because first the type needs to be added to a scope.