The main role can be extended.
Added Associable::rip_main_role method to be used for taking the
main role of a base role.

ClassType::complete_type can take over the main role of a base role, and
either use it as the class main role, or replacing a base of the class
main role. It also checks that the explicitly declared main role type is
derived from the old type. It also defers checking for full
implementations of the roles, because when a main role is inherited from
a base.

ClassType::allocate inherits VT entries from the base role that belongs
to its main role, and extends them with the abstract entries from the explicit
main role.

Role::allocate allows being called on ripped main roles, which are
chadow roles, but directly contained in a class.

Fixed InterfaceType::get_heritage not returning the base role correctly.

Added InterfaceType::lookup_inner method to handle forwarding of the
lookup into the base role.

InterfaceDefinitionValue can process a base role type name.