Added role aliasing.
Added Allocable::get_offset method to avoid messing with offset from
outside, and allow subclasses to override it.

Added Aliasing class, an almost dummy class to be returned by an
aliasing :Role declaration.

Added AS_ALIAS value to InheritAs enum.

Added Inheritable::alias method to turn a role into an alias for another.

Modified Role class so that an aliased role does not allocate data for
itself, and its VT entry will store the offset of the aliased role.

Added unqualify helper function.

Added :Role of: construct that takes the name of an existing role, to
be aliased.