Added :Global metatype, in order to replace :Singleton.
Added GlobalVariable subclass of Variable, storing a Ptr to
an interface, and implemented by a class.

RootScope can now register all the global variables.

Added Application interface, containing a single start method,
to be implemented by a global Main variable in the main module,
to be the new entry point.

Root now walks through the list of global variables, and creates the
entry function and the initializer and finalizer lists.

Factored out part of ClassPreinitializerValue to a preinitialize_class
function, to be used for global variable initializations, and functors.

Removed {,Partial}SingletonVariable, as methods can no longer be
defined in singletons.

Allowed initializer and finalizer functions without a declared body,
because these are always void.

Fixed TypeValue accessing the pivot type hint where there isn't any,
and not even needed.

Fixed TypeDefinitionValue::typize_typespec overwriting the typespec
of the current value object.

Added GlobalDefinitionValue to implement the :Global construct.

Singletons are only kept as much as necessary to use the colon
scope and the Std namespace.

The test no longer uses singletons for test cases.