The Std singleton is replaced by a global variable of Unit type.
Added GlobalNamespace subclass of GlobalVariable, to handle the
peculiarities of this case.

Fixed GlobalVariable methods not being virtual.

UnitType::where now returns explicit NOWHERE, like SingletonType did.

Std is now a GlobalNamespace instance with type <Std>, which is a
Unit subtype, so it won't be stored or passed as pivot argument.

Prefixing the main type of a regular global variable is moved to the
GlobalDefinitionValue, to allow GlobalVariable use any value type.