Added :Abstract metatype for identity type interfaces.
The :Interface metatype is now value type only.

Associable::make_shadow is now used to create a shadow of the this
object, because now shadows have the same type as the original, not
a type decided by the scope.

Relaxed some Associable methods not to throw exceptions, because
they're not needed in Implementation.

Moved IdentityType to class.cpp, and they're no longer a base type
for InterfaceType.

Added AbstractType, a sibling class of ClassType.

Added IdentityType::allocate_vt method to let AbstractType fill it
with dummy VT entries.

IdentityType::streamify now delegates the job to the streamifiable
implementation, like ImplementationConversionValue does, because
Streamifiable Ptr pivots are no longer possible.

InterfaceType got a few method copied from IdentityType, because
it is still cleaner this way.

Tweaked Implementation to work within identity types, by allocating
VT entries for the interface methods, without the VT headers, so
getting a Ptr to the Implementation is not possible, but overriding
the implementation methods is.

Added abstract_metatype.

The Application interface is now an Abstract.

Interface pivot types are no longer Ptr types.

The :Role control now always creates Role instances, not Implementation-s.