Fixed patching methods not using the correct role offset.
Since a patching method can't compute the offset during compile
time, because it depends on the classes derived from it, it must be
passed to it as a hidden argument during runtime.

Added VirtualEntry::compile method to allow VE-s compile some stuff
they need.

Added PatchMethodVirtualEntry class, which returns a label to a
privately compiled function trampoline that sets R11 to an offset value.

Added Associable::subst_offset to disallow direct access to the offset
member variable.

Replaced direct accesses to Associable offset-s.

Allocating a provisioned role now performs the patching of the provider
methods correctly. For every method that was overridden as a requirement
it finds the override role's offset, the provider role's offset, and
sets up a trampoline to store the difference to R11 as a virtual entry.

Added FunctionScope::associated_offset_storage, so functions can have
a hidden variable to store it when necessary.

Added Implementation::{allocate,relocate} methods that makes sure their
offset is 0, as no self adjustment is needed for implementation methods.

FunctionDefinitionValue generates the necessary code to process the
hidden R11 argument for patching methods.