Added stack accounting to determine per-function stack usage.
Added some methods to Asm64 to manage the stack accounting feature.

Modified Asm64 opcode generation to look out for stack related
opcodes with the accounting on, measuring changes in stack usage for
expected opcodes, and aborting on the rest.

Cleaned up StackOp and RegisterMemoryOp handling.

Added Once methods to schedule compilation of Value objects, used to
late compile the bodies of functors, instead of generating the code
inside the defining function body, then skipping it.

Modified Unwind::initiate method to mark and rewind stack accounting,
so that the destroying of temporary values on the stack in the unwind
branch are not counted multiple times, along with the normal control flow.

Surrounded the raise of some exceptions with an extra level of mark/rewind
in the cases where a temporary must be destroyed on the stack before
raising the exception itself.

The stack alignment hacks in Runtime::call_sysv{,_got} have the stack
accounting turned off, as it is almost impossible to programmatically
measure the stack change in these cases, so instead we manually record
a 24 bytes stack usage just to be safe.

Factored out a ControlValue::nonsense_result method from RaiseValue and
YieldValue, so that in the unlikely case of using such controls in a non-void
context still pretend to produce a result on the stack that pleases
the accounting as well.

Cleaned up TryValue handling yields from the try clause to be different
from the handling of exceptions from the fix clause.

FunctionDefinitionValue measures the stack usage of the function body,
so far just because.

Removed some obsolete equality matcher Value classes.