Stack relocation.
Added values/debug.cpp to contain the operations designed for debugging only,
currently FrameNameValue and DoubleStackValue.

Added FRAME_INFO_* constants to heap.h.

Added memcpy wrapper to the C runtime.

Runtime now stores the address and the isze of the task stack.

Split parts of lookup_frame_info function to caller_frame_info.

Moved the start function compilation from Root to Runtime, so it is
closer to the similar functions.

Added fix_stack function to follow the backtrace, and invoke the fixup
for each stack frame to relocate the ALIAS arguments of the functions.

Added double_stack function to relocate and double the task stack just because.

FunctionDefinitionValue now generates fixups for all functions to relocate
the ALIAS storage arguments if their value point to the stack.

Merged some strings in input.sh to make test runs more consistent.