Function call and evaluable cleanups.
CodeScopeValue no longer has a separate compile_evaluable method, instead
it has an am_evalue flag, which affects the normal compile method. Simple
values are now always returned in registers, even in evalue usage.

FunctionCallValue stores the original storage of the pivot argument, so if
it has to return it for void functions, it returns the same storage, not
just the address popped from the stack. Stack relative addresses are
not conflated with heap addresses anymore.

Fixed FunctionCallValue::unwind not handling specially the Code arguments.

RecordPostinitializerValue checks for the same storage that
RecordPreinitializerValue used.

Modified EvaluableValue to handle the result value according to the above
modifications, only managing the stack when necessary.