The ALIAS storage now stores and indirect offset as well.
Storage, when the where field is set to ALIAS, uses both the address
field to point to the alias location, RBP-relative, and the value field,
storing an indirect offset that must be applied after the address is
loaded into a register.

This makes it possible not to load an ALIAS into a register immediately
if  used as a container of another field. This way an address computed
from an ALIAS never needs to be spilled to a temporary alias, so all the
spilled ones can be guaranteed to point to the heap, and need no fixing
after a stack relocation.

Modified some ALIAS handling functions to take the indirect offset into
account. FunctionCallValue now adds it to the indirect address when
forwarding an ALIAS value to an ALIAS argument.

Variable allows its container to have an ALIAS storage, in which case
it just adjusts the indirect offset.

Added Value::hint_unalias method, used to let a Value know that it will
only be used as an rvalue, and should not return an ALIAS storage, as its
parent does not want to bother with register allocation just to deal with it.

VariableValue uses the hint_unalias method to allocate a register itself,
and return a MEMORY storage instead of an ALIAS, to be nice.

Unalias hinting uses Type::where to figure out which argument types would
need lvalue-like semantics, so it had to be added to some types just for
this purpose, like InterfaceType, AnyType, MultiType, UninitializedType.

Fixed MultiValue mixing up ALIAS with ALISTACK when used as the item
storage.