Made evaluables be backtrace-friendly by having their stack frame.
Added RetroScope subclass of CodeScope, it can be looked up, and it
allocates a space for an artificial return address plus parent frame
pointer.

The check_argument{,s} functions take a is_function_call flag to see
if they need to create a RetroScope instead of a CodeScope for arguments
in a Code context.

CodeScopeValue can tell of it needs retro operation by its scope,
and it no longer needs to juggle STACK result values, as the top of the
stack is no longer used for evalue bookkeeping.

A function call passing in a Code block compiles the necessary steps
to get its own frame pointer back, store its return address and parent frame
pointer to an alocated location, and use them to return to the evaluator.
This way the executing Code block has a frame pointer to its evaluator,
so the backtrace no longer skips it.

VariableValue is now aware of the potential surrounding retro scope, and
adjusts its RBP based address with the fixed offset of the retro space,
so references inside a Code block to variables declared outside of it still
work.