Optimizing unwind checks now depends on flags stored in Scope-s.
Remembering if a Value::unwind was ever called was fragile, as it depended
on the unwind method be called earlier, which may no longer be true with
parts of the code compiled deferredly. So we set some flags sooner.

Added Scope::be_unwindable method, for completeness.

Added CodeScope::be_unwindable method, to remember this state recursively.

Added FunctionScope::be_unwindable method, to terminate this chain.

RaisingDummy now calls the be_unwindable method of its outer scope,
so all enclosing scoped values will actually compile checks.

Moved most of CodeScopeValue compilation to compile_body method,
so that the following checks may be optimized out here, but not
in the RetroScopeValue subclass.

Removed the former optimizations from FunctionDefinitionValue, CodeScopeValue,
and YieldableValue.