Allowed :try to forward multivalue results from the body.
Added multicode_type to represent a Multi Code, which is actually
invalid, as Multi is not a value type. This allows the argument
checking wrap such multivalues in a code scope.

Added STACK to STACK nop store to Type::store.

Added missing MultiType::where case with AS_VALUE.

Added unpack method to TryValue and CodeScopeValue, so multivalue
results can be forwarded by these.

Multi unpacking evaluates the right hand side in a Multi context.

Removed the printing of the source file and line numbers in the test,
as these values will constantly change anyway.