Value::precompile returns clobbering stack or heap based variables.
Added extra bits and named constructors to the Regs bitset.

Added Allocable::lvalue_would_clobber method to return the Regs bitset
that would be clobbered if this variable was used as an lvalue argument.

Replaced unalias hinting with an lvalue argument type based setting.

VariableValue, ContainerIndexCalue, RbtreeIndexValue, and SliceIndexValue
now sets the bits for variable clobbering if used as an lvalue argument.

Renamed VariableValue::reg to unalias_reg to make it clear that this register
is allocated to be an address base. It is also allocated in all cases,
as it is impossible to decide in the precompile phase if it will be needed.

Rewrote VariableValue::compile a bit.

The BMEMORY storage is no longer used in practice.