Weak* container operations now use unified Value classes.
Replaced the Weak*Value classes with a common NosyRbtree*Value
classes, which get the rbtree elem type as a parameter, and figure
everything else out from it, so they work with all of WeakValueMap,
WeakIndexMap, and WeakSet.

NosytreeTemplateIdentifier also takes an elem type parameter to
keep life simple.

And NosytreeMemberValue, too.