Fixed VariableValue borrow handling.
Renamed Regs::aliasvars to relaxvars, as it is also used to prevent
value copies.

Fixed VariableValue setting the stack and heap flags for its child Value,
which would prevent them from making the proper reference adjustments.

Other Values that consider making a value copy now check for the
relaxvars flag, and don't do it if set.

Saved a reference flip-flop in RbtreeIndexValue.