First steps toward tuple typed type parameters.
The type system right now is a bit of a mess, as both the old and the
new approach is used simultaneously, but it will be cleaned up.

Added TupleMetaType class, the metatype of all (anonymous) tuple types.
It can also be instantiated with an anonymous initializer syntax,
which is invoked when braces are used without type context.

Added TupleType class for each tuple type created. Since they're
anonymous, eventually they will be cached by their keyword list,
but not now.

Added tuple0_type and tuple1_type type constants, to be used for zero
and one element tuples.

Code now requires a tuple type parameter instead of a value type.
TS constants were rewritten to contain a tuple parameter.

The typematch function handles conversion into tuple types. Conversion
to a zero element tuple is handled like conversion to Void, conversion
to a one element tuple is handled usually, and conversion to a multi
element tuple is not yet allowed.

The prefixing and unprefixing of typespecs with code_type became a bit
complex as tuples became required before prefixing, and unwanted after
unprefixing.

Cleaned up multivalue parsing a bit, but they are quickly becoming a
nuisance with their nonstandard handling, and may soon reach their end.