Split the handling of tuples without context.
Type-s with multiple type parameters expect to get a TypeTupleValue
pivot value.

The scopes of tuple types and the tuple metatype has a dummy pivot type,
because we can't yet represent tuple types with an arbitrary number
of parameters.

Added caching of TupleType instances, because these things are created on
the fly, and no scope actually owns them.

Added Unpacking identifier, because it is placed in the tuple metascope,
so the pivot type checking is practically impossible they way other
identifiers do it.

Moved the parsing of contextless tuples out of MultiValue into a new
typize_contextless_tuple function, and the resulting Value object is
chosen according to the result on this.

Replaced MultiValue class with LvalueTupleValue, which is now only responsible
of handling the case where all tuple elements are lvalues.

UnpackingValue now requires a LvalueTupleValue instance as the pivot value.

Added TypeTupleValue class to represent a tuple of type names.