Multiple retro arguments can be declared in theory.
Added RetroArgumentScope class to contain the retro variables in a
tuple Dvalue context.

Added RetroArgumentValue class to handle the RetroArgumentScope,
check the expressions inside to be bare declarations, and fixes them
based on the Dvalue type it has.

The check_argument function now also handles the Dvalue context specially,
adding a RetroArgumentScope before, and a RetroArgumentValue after the
typization of the argument expression.

Fixed check_arguments not noticing if too many arguments were given.

The typize function for tuple expressions in a type context now checks
the context, and creates a CodeBlockValue for {...} Code, and a
DataBlockValue for {...} Dvalue contexts, other contexts are disallowed.

From now on we explicitly disallow code blocks in nontrivial contexts.

Added DeclarationValue::fix_bare_retro method, because the fixing of bare
declaration for retros is a bit different (the variable cannot be made an
lvalue like with unpacking, as for retros it means something else).