Tuple values can be created in a tuple typed context.
Added TupleBlockValue class to assemble a literal tuple value,
and store the value on the stack.

A tuple value can be created from a literal tuple in a tuple typed
context (for the sake of :evaluate arguments), and in a tuple
Code typed context (for the sake of returning such a value).

Fixed EvaluateValue not adjusting the stack accounting if the
Code result is received on the stack.