Type parameters are looked up as regular arguments, not pivots.
In the typize function identifier names are checked for type names,
and for them the left subexpression is treated as the arguments.

Rewrote lookup function as lookup_identifier, and now instead of a single
Expr argument, it gets the Token, Args, Kwargs separately.

Type no longer checks the parameters in the match method, but returns
a TypeValue to do the job.

TypeValue no longer gets the full represented typespec, but does the
argument checking to collect the type parameters. It also provides
metatype contexts for each arguments, so a tuple type can be defined
with the anonymous initializer.

Initializers without a type context are an error again, no need to
treat them as tuple type definitions.

Tuples without context must be lvalue tuples.