The Iterator and Iterable types now take a tuple type parameter.
Added Anytuple and Sametuple types, and a bunch of related constants.

Allowed some TupleType::store operations, mostly out of lazyness,
so the existing iterator next implementations can continue returning
their values in REGISTER and MEMORY storage.

Iterator next methods now return single tuples instead of plain values,
and double tuples instead of Item-s (the memory layout is reversed).

Rewrote the test code to use a tuple of bare declarations as the
Dvalue argument of item iterations.

Array and Queue item iterator next methods push the tuple onto
the stack in the right order (index pushed first).

The :for control now pushes the next method result onto the stack
to avoid invalid tuple movements.

Added postprocess method to the RbtreeNext* classes. They are used for
Map item iterators to push the index and the value onto the stack
in the correct order.

Added hack to Function to allow the full result value be a tuple type.
This is needed for the Iterator nex method. So we allow a single
Sametuple type result value (illegal), which is replaced by the
actual matched tuple type.