Built-in iterators can return lvalues in tuples.
Normal functions are still not allowed to return lvalues, but in
the case of built-in iterators we can trust the next methods not to
mess with the underlying containers, so we let them return lvalues.

Fixed TupleType not considering that ALIAS storage elements can
be present.

Plain container iterators returning elements and items return an lvalue
to the element.

Map iterators returning items return an lvalue to the value.

Also fixed SliceNextItemValue using the wrong order of tuple elements.