Added TypeSpec argument to {Code,Retro}ScopeValue constructors.
This is so retro values can have a Code type, so passing them to
functions pushed and pops the correct storage.

Typematch removes the Code type from the source when converting to a
non-Code.

Typematch allows converting a single tuple to Any.

CodeBlockValue makes sure each statement's stack usage returns to the
starting value.