FunctionScope can adjust its frame pointer base offset.
Since many internal constructs and variables are RBP-relative, for
handling the frame offset in retro scopes it's not enough to adjust it in
the VariableValue-s, so FunctionScope can be adjusted instead while
compiling the bodies of retro scopes.

Added FunctionScope::adjust_frame_base_offset method.

Made some methods virtual for consistency.

Removed the retro adjustments from VariableValue.

Added tiny test case for this.