Rewrote VariableValue operation.
Removed the RELAXVARS flag from Regs, VariableValue no longer asks
its pivot value for special processing. That only helper in some rare
cases, and having a flag for it in Regs was a bit unclear.

NosytreeMemberValue gets its member type as a constructor argument,
as it may or may not be an lvalue, depending on the situation.

Fixed ArrayExtendValue using an lvalue for its argument and result type.

StringLiteralValue no longer uses BREGISTER for its result storage,
instead returns a MEMORY pointing to a static String value, as an
alternative to return a borrowable value.

Added sanity check to Value::need_lvalue to catch cases when non-lvalue
operations are asked for lvalue result.

VariableValue no longer checks in advance for single record or rvalue
record pivots, as this check was a bit questionable. Instead, single
records are checked by size comparisons, and rvalue records by STACK
storages. The latter is now more optimal for REGISTER member values, and
a bit less optimal for STACK ones.

Fixed VariableValue::need_lvalue asking for lvalue pivot even it that's
a Ptr.

VariableValue with a pivot value lets the pivot decide the borrowability
of the location.

Unborrowable pivots now return a value copy, this case is no longer
optimized, as the RELAXVARS flag was removed.

ALIAS storage is either returned as MEMORY, if borrowable, or as a value
copy.