Value copying or borrowing is handled by the new RvalueCastValue.
Whenever a value with an lvalue type is used where an rvalue is
expected, an RvalueCastValue instance is inserted. This handles
the unborrowable MEMORY/ALIAS/ALISTACK storages, and loads a
value copy instead.

It is currently disable for interface types, so arguments expecting
an interface will always keep the lvalue-ness. This is necessary for
ForEachValue to store an lvalue iterator.

Fixed String RecordWrapperIdentifier operations not having an lvalue
member type specified, which is now necessary to get is cast to rvalue
when needed.

Fixed RbtreeIndexValue not returning an lvalue result.

Modified RecordWrapperValue to also insert a RvalueCastValue if necessary.

Also InterpolationValue.

ContainedLvalue became much simpler, as most of its conversions are now
in RvalueCastValue.

VariableValue also became simpler.

Added Identifier::disable_rvalue_pivot_cast method. This is currently
a workaround for Variable, that has a rvalue pivot, so it works with
both rlavue and lvalue pivots, but should not have its pivot cast to
rvalue, as a lvalue pivot means lvalue variable as well.