Cleaned up FunctionCallValue argument processing.
Added PushedInfo struct to contain all necessary fields that are used for
each argument, and replaced the various pushed_* vectors with
pushed_infos.

The info now stores the original Storage, and it makes the stackfix,
borrow Storage-s and the pivot_alias_storage variable redundant. The
first two are now replaced by bool flags.