GenericLvalue related cleanups.
{Array,Queue,Container}IndexValue is finally not an Operation, not sure
why it was the first place. This was now necessary, as otherwise it
would contain two instances of the GenericLvalue base class.

GenericOperationValue is now a GenericLvalue, because assignment and
initialization returns lvalues.

CreateValue is no longer an explicit GenericLvalue, as its base
GenericOperationValue is it now.

Added some need_rvalue methods to GenericLvalue subclasses for clarity.

Moved pivot hint setting of VeriableValue from precompile to need_rvalue.

Fixed handling of uninitialized variable clobbering in VariableValue. In this case
we can't use either the pivot value's clobbers, or the variable's clobbers
(that works for assignment only).