Machine opcodes are virtual, translated by the X64 backend.
Split the instruction set to a virtual and an X64 one, although
currently they are the same, and the mappings are just type casts.

Split the asm64.* files to asm.* with the architecture independent
Asm base class, and to asm_x64.* with the Asm_X64 class.

Added Cc base class for translating virtual opcodes to actual ones.

Added Cc_X64 class that translates them to X64 opcodes using an
instance of Asm_X64.

X64 opcodes are now defined in the X namespace.

Added Asm::get_dc method to get the current data segment position.

Added Acc base class for Accounting, used to pass an instance to
Asm to be used for accounting.

The (unfortunately named) X64 class is now a subclass of Cc_X64.