X64 is now an abstract class.
The x64-specific implementation is now the horribly named X64_X64 class.
This inherits the implementation from Emu_X64.

Emu-related interfaces are now inherited virtually, so their implementation
can be inherited independently.

The data_reference method is now in Emu to require X64 implementations
to implement it.

Added data relocations to Emu_A64 to be able to implement this.

Added add_ref helper methods to Emu subclasses.

Runtime gets the source file names in the constructor, and compiles the
necessary data itself.