Cleaned up welcome/goodbye code.
They now include functionality for prologue/epilogue, only that they
store the stack frame on the system stack, and does not care about the
task stack. RBP is set to point to the system frame in welcome, and is
expected to point there before goodbye.

The FP is no longer used on A64 because the CFA is already generated to
help gdb to use RBP/X5 for the frame address.

A64 prologue and epilogue also uses pair stores and loads.